[CALCITE-2435] tests: refactor SqlTestFactory
