+++ core/src/test/java/org/apache/calcite/sql/test/SqlAdvisorTest.java
@@ -16,20 +16,12 @@
  */
 package org.apache.calcite.sql.test;
 
-import org.apache.calcite.rel.type.RelDataTypeFactory;
-import org.apache.calcite.rel.type.RelDataTypeSystem;
 import org.apache.calcite.sql.advise.SqlAdvisor;
 import org.apache.calcite.sql.advise.SqlAdvisorValidator;
 import org.apache.calcite.sql.advise.SqlSimpleParser;
-import org.apache.calcite.sql.fun.SqlStdOperatorTable;
 import org.apache.calcite.sql.parser.SqlParserUtil;
-import org.apache.calcite.sql.type.SqlTypeFactoryImpl;
-import org.apache.calcite.sql.validate.SqlConformance;
 import org.apache.calcite.sql.validate.SqlMoniker;
 import org.apache.calcite.sql.validate.SqlMonikerType;
-import org.apache.calcite.sql.validate.SqlValidator;
-import org.apache.calcite.sql.validate.SqlValidatorWithHints;
-import org.apache.calcite.test.MockCatalogReader;
 import org.apache.calcite.test.SqlValidatorTestCase;
 
 import org.junit.Assert;
@@ -52,6 +44,10 @@ import static org.junit.Assert.fail;
  * for SqlAdvisor.
  */
 public class SqlAdvisorTest extends SqlValidatorTestCase {
+  public static final SqlTestFactory ADVISOR_TEST_FACTORY = SqlTestFactory.INSTANCE.withValidator(
+      SqlAdvisorValidator::new
+  );
+
   //~ Static fields/initializers ---------------------------------------------
 
   private static final List<String> STAR_KEYWORD =
@@ -413,9 +409,7 @@ public class SqlAdvisorTest extends SqlValidatorTestCase {
   protected void assertHint(
       String sql,
       String expectedResults) throws Exception {
-    SqlValidatorWithHints validator =
-        (SqlValidatorWithHints) tester.getValidator();
-    SqlAdvisor advisor = tester.getFactory().createAdvisor(validator);
+    SqlAdvisor advisor = tester.getFactory().createAdvisor();
 
     SqlParserUtil.StringAndPos sap = SqlParserUtil.findPos(sql);
 
@@ -436,9 +430,7 @@ public class SqlAdvisorTest extends SqlValidatorTestCase {
    * @param expected Expected result after simplification.
    */
   protected void assertSimplify(String sql, String expected) {
-    SqlValidatorWithHints validator =
-        (SqlValidatorWithHints) tester.getValidator();
-    SqlAdvisor advisor = tester.getFactory().createAdvisor(validator);
+    SqlAdvisor advisor = tester.getFactory().createAdvisor();
 
     SqlParserUtil.StringAndPos sap = SqlParserUtil.findPos(sql);
     String actual = advisor.simplifySql(sap.sql, sap.cursor);
@@ -467,9 +459,7 @@ public class SqlAdvisorTest extends SqlValidatorTestCase {
       String sql,
       String expectedResults,
       String expectedWord) {
-    SqlValidatorWithHints validator =
-        (SqlValidatorWithHints) tester.getValidator();
-    SqlAdvisor advisor = tester.getFactory().createAdvisor(validator);
+    SqlAdvisor advisor = tester.getFactory().createAdvisor();
 
     SqlParserUtil.StringAndPos sap = SqlParserUtil.findPos(sql);
     final String[] replaced = {null};
@@ -527,7 +517,7 @@ public class SqlAdvisorTest extends SqlValidatorTestCase {
   }
 
   @Override public SqlTester getTester() {
-    return new SqlTesterImpl(new AdvisorTesterFactory());
+    return new SqlTesterImpl(ADVISOR_TEST_FACTORY);
   }
 
   /**
@@ -1233,29 +1223,6 @@ public class SqlAdvisorTest extends SqlValidatorTestCase {
     sql = "select 1 from emp group by ^ except select 2 from dept a";
     simplified = "SELECT * FROM emp GROUP BY _suggest_";
     assertSimplify(sql, simplified);
-  }
-
-  /** Factory that creates testers. */
-  private static class AdvisorTesterFactory extends DelegatingSqlTestFactory {
-    AdvisorTesterFactory() {
-      super(DefaultSqlTestFactory.INSTANCE);
-    }
-
-    @Override public SqlValidator getValidator(SqlTestFactory factory) {
-      final RelDataTypeFactory typeFactory =
-          new SqlTypeFactoryImpl(RelDataTypeSystem.DEFAULT);
-      final SqlConformance conformance = (SqlConformance) get("conformance");
-      final boolean caseSensitive = (Boolean) factory.get("caseSensitive");
-      return new SqlAdvisorValidator(
-          SqlStdOperatorTable.instance(),
-          new MockCatalogReader(typeFactory, caseSensitive).init(),
-          typeFactory,
-          conformance);
-    }
-
-    @Override public SqlAdvisor createAdvisor(SqlValidatorWithHints validator) {
-      return new SqlAdvisor(validator);
-    }
   }
 }
