+++ core/src/test/java/org/apache/calcite/test/catalog/CompoundNameColumn.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to you under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.calcite.test.catalog;
+
+import org.apache.calcite.rel.type.RelDataType;
+
+/** Column having names with multiple parts. */
+final class CompoundNameColumn {
+  final String first;
+  final String second;
+  final RelDataType type;
+
+  CompoundNameColumn(String first, String second, RelDataType type) {
+    this.first = first;
+    this.second = second;
+    this.type = type;
+  }
+
+  String getName() {
+    return (first.isEmpty() ? "" : ("\"" + first + "\"."))
+        + ("\"" + second + "\"");
+  }
+}
+
+// End CompoundNameColumn.java
