+++ core/src/test/java/org/apache/calcite/test/catalog/MockCatalogReader.java
@@ -14,13 +14,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.calcite.test;
+package org.apache.calcite.test.catalog;
 
 import org.apache.calcite.adapter.java.JavaTypeFactory;
 import org.apache.calcite.config.CalciteConnectionConfig;
 import org.apache.calcite.jdbc.CalcitePrepare;
 import org.apache.calcite.jdbc.CalciteSchema;
-import org.apache.calcite.linq4j.Ord;
 import org.apache.calcite.linq4j.QueryProvider;
 import org.apache.calcite.linq4j.Queryable;
 import org.apache.calcite.linq4j.tree.Expression;
@@ -44,7 +43,6 @@ import org.apache.calcite.rel.type.RelDataTypeComparability;
 import org.apache.calcite.rel.type.RelDataTypeFactory;
 import org.apache.calcite.rel.type.RelDataTypeFamily;
 import org.apache.calcite.rel.type.RelDataTypeField;
-import org.apache.calcite.rel.type.RelDataTypeFieldImpl;
 import org.apache.calcite.rel.type.RelDataTypeImpl;
 import org.apache.calcite.rel.type.RelDataTypePrecedenceList;
 import org.apache.calcite.rel.type.RelProtoDataType;
@@ -53,8 +51,6 @@ import org.apache.calcite.rel.type.StructKind;
 import org.apache.calcite.rex.RexBuilder;
 import org.apache.calcite.rex.RexInputRef;
 import org.apache.calcite.rex.RexNode;
-import org.apache.calcite.rex.RexUtil;
-import org.apache.calcite.schema.ColumnStrategy;
 import org.apache.calcite.schema.CustomColumnResolvingTable;
 import org.apache.calcite.schema.ExtensibleTable;
 import org.apache.calcite.schema.Path;
@@ -64,8 +60,6 @@ import org.apache.calcite.schema.Schemas;
 import org.apache.calcite.schema.Statistic;
 import org.apache.calcite.schema.StreamableTable;
 import org.apache.calcite.schema.Table;
-import org.apache.calcite.schema.TableMacro;
-import org.apache.calcite.schema.TranslatableTable;
 import org.apache.calcite.schema.Wrapper;
 import org.apache.calcite.schema.impl.AbstractSchema;
 import org.apache.calcite.schema.impl.ModifiableViewTable;
@@ -73,14 +67,10 @@ import org.apache.calcite.schema.impl.ViewTableMacro;
 import org.apache.calcite.sql.SqlAccessType;
 import org.apache.calcite.sql.SqlCall;
 import org.apache.calcite.sql.SqlCollation;
-import org.apache.calcite.sql.SqlFunction;
 import org.apache.calcite.sql.SqlIdentifier;
 import org.apache.calcite.sql.SqlIntervalQualifier;
 import org.apache.calcite.sql.SqlKind;
 import org.apache.calcite.sql.SqlNode;
-import org.apache.calcite.sql.fun.SqlStdOperatorTable;
-import org.apache.calcite.sql.parser.SqlParserPos;
-import org.apache.calcite.sql.type.ObjectSqlType;
 import org.apache.calcite.sql.type.SqlTypeName;
 import org.apache.calcite.sql.validate.SqlModality;
 import org.apache.calcite.sql.validate.SqlMonotonicity;
@@ -88,31 +78,31 @@ import org.apache.calcite.sql.validate.SqlNameMatcher;
 import org.apache.calcite.sql.validate.SqlNameMatchers;
 import org.apache.calcite.sql.validate.SqlValidatorCatalogReader;
 import org.apache.calcite.sql.validate.SqlValidatorUtil;
-import org.apache.calcite.sql2rel.InitializerContext;
 import org.apache.calcite.sql2rel.InitializerExpressionFactory;
 import org.apache.calcite.sql2rel.NullInitializerExpressionFactory;
+import org.apache.calcite.test.JdbcTest;
 import org.apache.calcite.util.ImmutableBitSet;
 import org.apache.calcite.util.ImmutableIntList;
-import org.apache.calcite.util.Litmus;
 import org.apache.calcite.util.Pair;
 import org.apache.calcite.util.Util;
 
+import com.google.common.cache.CacheBuilder;
+import com.google.common.cache.CacheLoader;
+import com.google.common.cache.LoadingCache;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
 
+import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Type;
-import java.math.BigDecimal;
 import java.nio.charset.Charset;
 import java.util.AbstractList;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Mock implementation of {@link SqlValidatorCatalogReader} which returns tables
@@ -120,16 +110,39 @@ import java.util.concurrent.atomic.AtomicInteger;
  * Also two streams "ORDERS", "SHIPMENTS";
  * and a view "EMP_20".
  */
-public class MockCatalogReader extends CalciteCatalogReader {
+public abstract class MockCatalogReader extends CalciteCatalogReader {
   //~ Static fields/initializers ---------------------------------------------
 
   static final String DEFAULT_CATALOG = "CATALOG";
   static final String DEFAULT_SCHEMA = "SALES";
   static final List<String> PREFIX = ImmutableList.of(DEFAULT_SCHEMA);
 
-  //~ Instance fields --------------------------------------------------------
+  private static final LoadingCache<Class<? extends MockCatalogReader>,
+      LoadingCache<Pair<RelDataTypeFactory, Boolean>, Prepare.CatalogReader>> CACHE =
+      CacheBuilder.newBuilder().build(
+          CacheLoader.from((clazz) -> {
+            Constructor<? extends MockCatalogReader> constructor;
+            try {
+              constructor = clazz.getConstructor(RelDataTypeFactory.class, boolean.class);
+            } catch (NoSuchMethodException e) {
+              throw new IllegalArgumentException(
+                  "There should be (RelDataTypeFactory, boolean) constructor", e);
+            }
 
-  private RelDataType addressType;
+            return CacheBuilder.newBuilder().build(
+                CacheLoader.from((key) -> createCatalogReader(constructor, key.left, key.right)));
+
+          })
+      );
+
+  public static Prepare.CatalogReader create(
+      Class<? extends MockCatalogReader> clazz,
+      RelDataTypeFactory typeFactory,
+      boolean caseSensitive) {
+    return CACHE.getUnchecked(clazz).getUnchecked(new Pair<>(typeFactory, caseSensitive));
+  }
+
+  //~ Instance fields --------------------------------------------------------
 
   //~ Constructors -----------------------------------------------------------
 
@@ -148,6 +161,17 @@ public class MockCatalogReader extends CalciteCatalogReader {
         typeFactory, null);
   }
 
+  private static Prepare.CatalogReader createCatalogReader(
+      Constructor<? extends MockCatalogReader> constructor,
+      RelDataTypeFactory typeFactory,
+      boolean caseSensitive) {
+    try {
+      return constructor.newInstance(typeFactory, caseSensitive).init();
+    } catch (InstantiationException | IllegalAccessException | InvocationTargetException e) {
+      throw new IllegalStateException("Unable to instantiate " + constructor, e);
+    }
+  }
+
   @Override public boolean isCaseSensitive() {
     return nameMatcher.isCaseSensitive();
   }
@@ -159,359 +183,9 @@ public class MockCatalogReader extends CalciteCatalogReader {
   /**
    * Initializes this catalog reader.
    */
-  public MockCatalogReader init() {
-    final Fixture f = new Fixture();
-    addressType = f.addressType;
+  public abstract MockCatalogReader init();
 
-    // Register "SALES" schema.
-    MockSchema salesSchema = new MockSchema("SALES");
-    registerSchema(salesSchema);
-
-    // Register "EMP" table with customer InitializerExpressionFactory
-    // to check whether newDefaultValue method called or not.
-    final InitializerExpressionFactory countingInitializerExpressionFactory =
-        new CountingFactory(ImmutableList.of("DEPTNO"));
-
-    registerType(
-        ImmutableList.of(salesSchema.getCatalogName(), salesSchema.getName(),
-            "customBigInt"),
-        typeFactory -> typeFactory.createSqlType(SqlTypeName.BIGINT));
-
-    // Register "EMP" table.
-    final MockTable empTable =
-        MockTable.create(this, salesSchema, "EMP", false, 14, null,
-            countingInitializerExpressionFactory);
-    empTable.addColumn("EMPNO", f.intType, true);
-    empTable.addColumn("ENAME", f.varchar20Type);
-    empTable.addColumn("JOB", f.varchar10Type);
-    empTable.addColumn("MGR", f.intTypeNull);
-    empTable.addColumn("HIREDATE", f.timestampType);
-    empTable.addColumn("SAL", f.intType);
-    empTable.addColumn("COMM", f.intType);
-    empTable.addColumn("DEPTNO", f.intType);
-    empTable.addColumn("SLACKER", f.booleanType);
-    registerTable(empTable);
-
-    // Register "EMPNULLABLES" table with nullable columns.
-    final MockTable empNullablesTable =
-        MockTable.create(this, salesSchema, "EMPNULLABLES", false, 14);
-    empNullablesTable.addColumn("EMPNO", f.intType, true);
-    empNullablesTable.addColumn("ENAME", f.varchar20Type);
-    empNullablesTable.addColumn("JOB", f.varchar10TypeNull);
-    empNullablesTable.addColumn("MGR", f.intTypeNull);
-    empNullablesTable.addColumn("HIREDATE", f.timestampTypeNull);
-    empNullablesTable.addColumn("SAL", f.intTypeNull);
-    empNullablesTable.addColumn("COMM", f.intTypeNull);
-    empNullablesTable.addColumn("DEPTNO", f.intTypeNull);
-    empNullablesTable.addColumn("SLACKER", f.booleanTypeNull);
-    registerTable(empNullablesTable);
-
-    // Register "EMPDEFAULTS" table with default values for some columns.
-    final MockTable empDefaultsTable =
-        MockTable.create(this, salesSchema, "EMPDEFAULTS", false, 14, null,
-            new EmpInitializerExpressionFactory());
-    empDefaultsTable.addColumn("EMPNO", f.intType, true);
-    empDefaultsTable.addColumn("ENAME", f.varchar20Type);
-    empDefaultsTable.addColumn("JOB", f.varchar10TypeNull);
-    empDefaultsTable.addColumn("MGR", f.intTypeNull);
-    empDefaultsTable.addColumn("HIREDATE", f.timestampTypeNull);
-    empDefaultsTable.addColumn("SAL", f.intTypeNull);
-    empDefaultsTable.addColumn("COMM", f.intTypeNull);
-    empDefaultsTable.addColumn("DEPTNO", f.intTypeNull);
-    empDefaultsTable.addColumn("SLACKER", f.booleanTypeNull);
-    registerTable(empDefaultsTable);
-
-    // Register "EMP_B" table. As "EMP", birth with a "BIRTHDATE" column.
-    final MockTable empBTable =
-        MockTable.create(this, salesSchema, "EMP_B", false, 14);
-    empBTable.addColumn("EMPNO", f.intType, true);
-    empBTable.addColumn("ENAME", f.varchar20Type);
-    empBTable.addColumn("JOB", f.varchar10Type);
-    empBTable.addColumn("MGR", f.intTypeNull);
-    empBTable.addColumn("HIREDATE", f.timestampType);
-    empBTable.addColumn("SAL", f.intType);
-    empBTable.addColumn("COMM", f.intType);
-    empBTable.addColumn("DEPTNO", f.intType);
-    empBTable.addColumn("SLACKER", f.booleanType);
-    empBTable.addColumn("BIRTHDATE", f.dateType);
-    registerTable(empBTable);
-
-    // Register "DEPT" table.
-    MockTable deptTable = MockTable.create(this, salesSchema, "DEPT", false, 4);
-    deptTable.addColumn("DEPTNO", f.intType, true);
-    deptTable.addColumn("NAME", f.varchar10Type);
-    registerTable(deptTable);
-
-    // Register "DEPT_NESTED" table.
-    MockTable deptNestedTable =
-        MockTable.create(this, salesSchema, "DEPT_NESTED", false, 4);
-    deptNestedTable.addColumn("DEPTNO", f.intType, true);
-    deptNestedTable.addColumn("NAME", f.varchar10Type);
-    deptNestedTable.addColumn("SKILL", f.skillRecordType);
-    deptNestedTable.addColumn("EMPLOYEES", f.empListType);
-    registerTable(deptNestedTable);
-
-    // Register "BONUS" table.
-    MockTable bonusTable =
-        MockTable.create(this, salesSchema, "BONUS", false, 0);
-    bonusTable.addColumn("ENAME", f.varchar20Type);
-    bonusTable.addColumn("JOB", f.varchar10Type);
-    bonusTable.addColumn("SAL", f.intType);
-    bonusTable.addColumn("COMM", f.intType);
-    registerTable(bonusTable);
-
-    // Register "SALGRADE" table.
-    MockTable salgradeTable =
-        MockTable.create(this, salesSchema, "SALGRADE", false, 5);
-    salgradeTable.addColumn("GRADE", f.intType, true);
-    salgradeTable.addColumn("LOSAL", f.intType);
-    salgradeTable.addColumn("HISAL", f.intType);
-    registerTable(salgradeTable);
-
-    // Register "EMP_ADDRESS" table
-    MockTable contactAddressTable =
-        MockTable.create(this, salesSchema, "EMP_ADDRESS", false, 26);
-    contactAddressTable.addColumn("EMPNO", f.intType, true);
-    contactAddressTable.addColumn("HOME_ADDRESS", addressType);
-    contactAddressTable.addColumn("MAILING_ADDRESS", addressType);
-    registerTable(contactAddressTable);
-
-    // Register "DYNAMIC" schema.
-    MockSchema dynamicSchema = new MockSchema("DYNAMIC");
-    registerSchema(dynamicSchema);
-
-    MockTable nationTable =
-        new MockDynamicTable(this, dynamicSchema.getCatalogName(),
-            dynamicSchema.getName(), "NATION", false, 100);
-    registerTable(nationTable);
-
-    MockTable customerTable =
-        new MockDynamicTable(this, dynamicSchema.getCatalogName(),
-            dynamicSchema.getName(), "CUSTOMER", false, 100);
-    registerTable(customerTable);
-
-    // Register "CUSTOMER" schema.
-    MockSchema customerSchema = new MockSchema("CUSTOMER");
-    registerSchema(customerSchema);
-
-    // Register "CONTACT" table.
-    MockTable contactTable = MockTable.create(this, customerSchema, "CONTACT",
-        false, 1000);
-    contactTable.addColumn("CONTACTNO", f.intType);
-    contactTable.addColumn("FNAME", f.varchar10Type);
-    contactTable.addColumn("LNAME", f.varchar10Type);
-    contactTable.addColumn("EMAIL", f.varchar20Type);
-    contactTable.addColumn("COORD", f.rectilinearCoordType);
-    registerTable(contactTable);
-
-    // Register "CONTACT_PEEK" table. The
-    MockTable contactPeekTable =
-        MockTable.create(this, customerSchema, "CONTACT_PEEK", false, 1000);
-    contactPeekTable.addColumn("CONTACTNO", f.intType);
-    contactPeekTable.addColumn("FNAME", f.varchar10Type);
-    contactPeekTable.addColumn("LNAME", f.varchar10Type);
-    contactPeekTable.addColumn("EMAIL", f.varchar20Type);
-    contactPeekTable.addColumn("COORD", f.rectilinearPeekCoordType);
-    contactPeekTable.addColumn("COORD_NE", f.rectilinearPeekNoExpandCoordType);
-    registerTable(contactPeekTable);
-
-    // Register "ACCOUNT" table.
-    MockTable accountTable = MockTable.create(this, customerSchema, "ACCOUNT",
-        false, 457);
-    accountTable.addColumn("ACCTNO", f.intType);
-    accountTable.addColumn("TYPE", f.varchar20Type);
-    accountTable.addColumn("BALANCE", f.intType);
-    registerTable(accountTable);
-
-    // Register "ORDERS" stream.
-    MockTable ordersStream = MockTable.create(this, salesSchema, "ORDERS",
-        true, Double.POSITIVE_INFINITY);
-    ordersStream.addColumn("ROWTIME", f.timestampType);
-    ordersStream.addMonotonic("ROWTIME");
-    ordersStream.addColumn("PRODUCTID", f.intType);
-    ordersStream.addColumn("ORDERID", f.intType);
-    registerTable(ordersStream);
-
-    // Register "SHIPMENTS" stream.
-    // "ROWTIME" is not column 0, just to mix things up.
-    MockTable shipmentsStream = MockTable.create(this, salesSchema, "SHIPMENTS",
-        true, Double.POSITIVE_INFINITY);
-    shipmentsStream.addColumn("ORDERID", f.intType);
-    shipmentsStream.addColumn("ROWTIME", f.timestampType);
-    shipmentsStream.addMonotonic("ROWTIME");
-    registerTable(shipmentsStream);
-
-    // Register "PRODUCTS" table.
-    MockTable productsTable = MockTable.create(this, salesSchema, "PRODUCTS",
-        false, 200D);
-    productsTable.addColumn("PRODUCTID", f.intType);
-    productsTable.addColumn("NAME", f.varchar20Type);
-    productsTable.addColumn("SUPPLIERID", f.intType);
-    registerTable(productsTable);
-
-    // Register "SUPPLIERS" table.
-    MockTable suppliersTable = MockTable.create(this, salesSchema, "SUPPLIERS",
-        false, 10D);
-    suppliersTable.addColumn("SUPPLIERID", f.intType);
-    suppliersTable.addColumn("NAME", f.varchar20Type);
-    suppliersTable.addColumn("CITY", f.intType);
-    registerTable(suppliersTable);
-
-    // Register "EMP_20" and "EMPNULLABLES_20 views.
-    // Same columns as "EMP" amd "EMPNULLABLES",
-    // but "DEPTNO" not visible and set to 20 by default
-    // and "SAL" is visible but must be greater than 1000,
-    // which is the equivalent of:
-    //   SELECT EMPNO, ENAME, JOB, MGR, HIREDATE, SAL, COMM, SLACKER
-    //   FROM EMP
-    //   WHERE DEPTNO = 20 AND SAL > 1000
-    final ImmutableIntList m0 = ImmutableIntList.of(0, 1, 2, 3, 4, 5, 6, 8);
-    MockTable emp20View =
-        new MockViewTable(this, salesSchema.getCatalogName(), salesSchema.name,
-            "EMP_20", false, 600, empTable, m0, null,
-            NullInitializerExpressionFactory.INSTANCE) {
-          public RexNode getConstraint(RexBuilder rexBuilder,
-              RelDataType tableRowType) {
-            final RelDataTypeField deptnoField =
-                tableRowType.getFieldList().get(7);
-            final RelDataTypeField salField =
-                tableRowType.getFieldList().get(5);
-            final List<RexNode> nodes = Arrays.asList(
-                rexBuilder.makeCall(SqlStdOperatorTable.EQUALS,
-                    rexBuilder.makeInputRef(deptnoField.getType(),
-                        deptnoField.getIndex()),
-                    rexBuilder.makeExactLiteral(BigDecimal.valueOf(20L),
-                        deptnoField.getType())),
-                rexBuilder.makeCall(SqlStdOperatorTable.GREATER_THAN,
-                    rexBuilder.makeInputRef(salField.getType(),
-                        salField.getIndex()),
-                    rexBuilder.makeExactLiteral(BigDecimal.valueOf(1000L),
-                        salField.getType())));
-            return RexUtil.composeConjunction(rexBuilder, nodes, false);
-          }
-        };
-    salesSchema.addTable(Util.last(emp20View.getQualifiedName()));
-    emp20View.addColumn("EMPNO", f.intType);
-    emp20View.addColumn("ENAME", f.varchar20Type);
-    emp20View.addColumn("JOB", f.varchar10Type);
-    emp20View.addColumn("MGR", f.intTypeNull);
-    emp20View.addColumn("HIREDATE", f.timestampType);
-    emp20View.addColumn("SAL", f.intType);
-    emp20View.addColumn("COMM", f.intType);
-    emp20View.addColumn("SLACKER", f.booleanType);
-    registerTable(emp20View);
-
-    MockTable empNullables20View =
-        new MockViewTable(this, salesSchema.getCatalogName(), salesSchema.name,
-            "EMPNULLABLES_20", false, 600, empNullablesTable, m0, null,
-            NullInitializerExpressionFactory.INSTANCE) {
-          public RexNode getConstraint(RexBuilder rexBuilder,
-              RelDataType tableRowType) {
-            final RelDataTypeField deptnoField =
-                tableRowType.getFieldList().get(7);
-            final RelDataTypeField salField =
-                tableRowType.getFieldList().get(5);
-            final List<RexNode> nodes = Arrays.asList(
-                rexBuilder.makeCall(SqlStdOperatorTable.EQUALS,
-                    rexBuilder.makeInputRef(deptnoField.getType(),
-                        deptnoField.getIndex()),
-                    rexBuilder.makeExactLiteral(BigDecimal.valueOf(20L),
-                        deptnoField.getType())),
-                rexBuilder.makeCall(SqlStdOperatorTable.GREATER_THAN,
-                    rexBuilder.makeInputRef(salField.getType(),
-                        salField.getIndex()),
-                    rexBuilder.makeExactLiteral(BigDecimal.valueOf(1000L),
-                        salField.getType())));
-            return RexUtil.composeConjunction(rexBuilder, nodes, false);
-          }
-        };
-    salesSchema.addTable(Util.last(empNullables20View.getQualifiedName()));
-    empNullables20View.addColumn("EMPNO", f.intType);
-    empNullables20View.addColumn("ENAME", f.varchar20Type);
-    empNullables20View.addColumn("JOB", f.varchar10TypeNull);
-    empNullables20View.addColumn("MGR", f.intTypeNull);
-    empNullables20View.addColumn("HIREDATE", f.timestampTypeNull);
-    empNullables20View.addColumn("SAL", f.intTypeNull);
-    empNullables20View.addColumn("COMM", f.intTypeNull);
-    empNullables20View.addColumn("SLACKER", f.booleanTypeNull);
-    registerTable(empNullables20View);
-
-    MockSchema structTypeSchema = new MockSchema("STRUCT");
-    registerSchema(structTypeSchema);
-    final List<CompoundNameColumn> columns = Arrays.asList(
-        new CompoundNameColumn("", "K0", f.varchar20Type),
-        new CompoundNameColumn("", "C1", f.varchar20Type),
-        new CompoundNameColumn("F1", "A0", f.intType),
-        new CompoundNameColumn("F2", "A0", f.booleanType),
-        new CompoundNameColumn("F0", "C0", f.intType),
-        new CompoundNameColumn("F1", "C0", f.intTypeNull),
-        new CompoundNameColumn("F0", "C1", f.intType),
-        new CompoundNameColumn("F1", "C2", f.intType),
-        new CompoundNameColumn("F2", "C3", f.intType));
-    final CompoundNameColumnResolver structTypeTableResolver =
-        new CompoundNameColumnResolver(columns, "F0");
-    final MockTable structTypeTable =
-        MockTable.create(this, structTypeSchema, "T", false, 100,
-            structTypeTableResolver);
-    for (CompoundNameColumn column : columns) {
-      structTypeTable.addColumn(column.getName(), column.type);
-    }
-    registerTable(structTypeTable);
-
-    final List<CompoundNameColumn> columnsNullable = Arrays.asList(
-        new CompoundNameColumn("", "K0", f.varchar20TypeNull),
-        new CompoundNameColumn("", "C1", f.varchar20TypeNull),
-        new CompoundNameColumn("F1", "A0", f.intTypeNull),
-        new CompoundNameColumn("F2", "A0", f.booleanTypeNull),
-        new CompoundNameColumn("F0", "C0", f.intTypeNull),
-        new CompoundNameColumn("F1", "C0", f.intTypeNull),
-        new CompoundNameColumn("F0", "C1", f.intTypeNull),
-        new CompoundNameColumn("F1", "C2", f.intType),
-        new CompoundNameColumn("F2", "C3", f.intTypeNull));
-    final MockTable structNullableTypeTable =
-        MockTable.create(this, structTypeSchema, "T_NULLABLES", false, 100,
-            structTypeTableResolver);
-    for (CompoundNameColumn column : columnsNullable) {
-      structNullableTypeTable.addColumn(column.getName(), column.type);
-    }
-    registerTable(structNullableTypeTable);
-
-    // Register "STRUCT.T_10" view.
-    // Same columns as "STRUCT.T",
-    // but "F0.C0" is set to 10 by default,
-    // which is the equivalent of:
-    //   SELECT *
-    //   FROM T
-    //   WHERE F0.C0 = 10
-    // This table uses MockViewTable which does not populate the constrained columns with default
-    // values on INSERT.
-    final ImmutableIntList m1 = ImmutableIntList.of(0, 1, 2, 3, 4, 5, 6, 7, 8);
-    MockTable struct10View =
-        new MockViewTable(this, structTypeSchema.getCatalogName(),
-            structTypeSchema.name, "T_10", false, 20, structTypeTable,
-            m1, structTypeTableResolver,
-            NullInitializerExpressionFactory.INSTANCE) {
-          @Override public RexNode getConstraint(RexBuilder rexBuilder,
-              RelDataType tableRowType) {
-            final RelDataTypeField c0Field =
-                tableRowType.getFieldList().get(4);
-            return rexBuilder.makeCall(SqlStdOperatorTable.EQUALS,
-                rexBuilder.makeInputRef(c0Field.getType(),
-                    c0Field.getIndex()),
-                rexBuilder.makeExactLiteral(BigDecimal.valueOf(10L),
-                    c0Field.getType()));
-          }
-        };
-    structTypeSchema.addTable(Util.last(struct10View.getQualifiedName()));
-    for (CompoundNameColumn column : columns) {
-      struct10View.addColumn(column.getName(), column.type);
-    }
-    registerTable(struct10View);
-    registerTablesWithRollUp(salesSchema, f);
-    return this;
-  }
-
-  private void registerTablesWithRollUp(MockSchema schema, Fixture f) {
+  protected void registerTablesWithRollUp(MockSchema schema, Fixture f) {
     // Register "EMP_R" table. Contains a rolled up column.
     final MockTable empRolledTable =
             MockTable.create(this, schema, "EMP_R", false, 14);
@@ -543,80 +217,6 @@ public class MockCatalogReader extends CalciteCatalogReader {
     nestedEmpRolledTable.addColumn("SLACKINGMIN", f.intType);
     nestedEmpRolledTable.registerRolledUpColumn("SLACKINGMIN");
     registerTable(nestedEmpRolledTable);
-  }
-
-  /** Adds some extra tables to the mock catalog. These increase the time and
-   * complexity of initializing the catalog (because they contain views whose
-   * SQL needs to be parsed) and so are not used for all tests. */
-  public MockCatalogReader init2() {
-    MockSchema salesSchema = new MockSchema("SALES");
-    // Same as "EMP_20" except it uses ModifiableViewTable which populates
-    // constrained columns with default values on INSERT and has a single constraint on DEPTNO.
-    List<String> empModifiableViewNames = ImmutableList.of(
-        salesSchema.getCatalogName(), salesSchema.name, "EMP_MODIFIABLEVIEW");
-    TableMacro empModifiableViewMacro = MockModifiableViewRelOptTable.viewMacro(rootSchema,
-        "select EMPNO, ENAME, JOB, MGR, HIREDATE, SAL, COMM, SLACKER from EMPDEFAULTS"
-            + " where DEPTNO = 20", empModifiableViewNames.subList(0, 2),
-        ImmutableList.of(empModifiableViewNames.get(2)), true);
-    TranslatableTable empModifiableView = empModifiableViewMacro.apply(ImmutableList.of());
-    MockModifiableViewRelOptTable mockEmpViewTable = MockModifiableViewRelOptTable.create(
-        (MockModifiableViewRelOptTable.MockModifiableViewTable) empModifiableView, this,
-        empModifiableViewNames.get(0), empModifiableViewNames.get(1),
-        empModifiableViewNames.get(2), false, 20, null);
-    registerTable(mockEmpViewTable);
-
-    // Same as "EMP_MODIFIABLEVIEW" except that all columns are in the view, columns are reordered,
-    // and there is an `extra` extended column.
-    List<String> empModifiableViewNames2 = ImmutableList.of(
-        salesSchema.getCatalogName(), salesSchema.name, "EMP_MODIFIABLEVIEW2");
-    TableMacro empModifiableViewMacro2 = MockModifiableViewRelOptTable.viewMacro(rootSchema,
-        "select ENAME, EMPNO, JOB, DEPTNO, SLACKER, SAL, EXTRA, HIREDATE, MGR, COMM"
-            + " from EMPDEFAULTS extend (EXTRA boolean)"
-            + " where DEPTNO = 20", empModifiableViewNames2.subList(0, 2),
-        ImmutableList.of(empModifiableViewNames.get(2)), true);
-    TranslatableTable empModifiableView2 = empModifiableViewMacro2.apply(ImmutableList.of());
-    MockModifiableViewRelOptTable mockEmpViewTable2 = MockModifiableViewRelOptTable.create(
-        (MockModifiableViewRelOptTable.MockModifiableViewTable) empModifiableView2, this,
-        empModifiableViewNames2.get(0), empModifiableViewNames2.get(1),
-        empModifiableViewNames2.get(2), false, 20, null);
-    registerTable(mockEmpViewTable2);
-
-    // Same as "EMP_MODIFIABLEVIEW" except that comm is not in the view.
-    List<String> empModifiableViewNames3 = ImmutableList.of(
-        salesSchema.getCatalogName(), salesSchema.name, "EMP_MODIFIABLEVIEW3");
-    TableMacro empModifiableViewMacro3 = MockModifiableViewRelOptTable.viewMacro(rootSchema,
-        "select EMPNO, ENAME, JOB, MGR, HIREDATE, SAL, SLACKER from EMPDEFAULTS"
-            + " where DEPTNO = 20", empModifiableViewNames3.subList(0, 2),
-        ImmutableList.of(empModifiableViewNames3.get(2)), true);
-    TranslatableTable empModifiableView3 = empModifiableViewMacro3.apply(ImmutableList.of());
-    MockModifiableViewRelOptTable mockEmpViewTable3 = MockModifiableViewRelOptTable.create(
-        (MockModifiableViewRelOptTable.MockModifiableViewTable) empModifiableView3, this,
-        empModifiableViewNames3.get(0), empModifiableViewNames3.get(1),
-        empModifiableViewNames3.get(2), false, 20, null);
-    registerTable(mockEmpViewTable3);
-
-    MockSchema structTypeSchema = new MockSchema("STRUCT");
-    registerSchema(structTypeSchema);
-    final Fixture f = new Fixture();
-    final List<CompoundNameColumn> columnsExtended = Arrays.asList(
-        new CompoundNameColumn("", "K0", f.varchar20TypeNull),
-        new CompoundNameColumn("", "C1", f.varchar20TypeNull),
-        new CompoundNameColumn("F0", "C0", f.intType),
-        new CompoundNameColumn("F1", "C1", f.intTypeNull));
-    final List<CompoundNameColumn> extendedColumns =
-        new ArrayList<CompoundNameColumn>(columnsExtended);
-    extendedColumns.add(new CompoundNameColumn("F2", "C2", f.varchar20Type));
-    final CompoundNameColumnResolver structExtendedTableResolver =
-        new CompoundNameColumnResolver(extendedColumns, "F0");
-    final MockTable structExtendedTypeTable =
-        MockTable.create(this, structTypeSchema, "T_EXTEND", false, 100,
-            structExtendedTableResolver);
-    for (CompoundNameColumn column : columnsExtended) {
-      structExtendedTypeTable.addColumn(column.getName(), column.type);
-    }
-    registerTable(structExtendedTypeTable);
-
-    return this;
   }
 
   //~ Methods ----------------------------------------------------------------
@@ -660,14 +260,6 @@ public class MockCatalogReader extends CalciteCatalogReader {
   private void registerNestedSchema(MockSchema parentSchema, MockSchema schema) {
     rootSchema.getSubSchema(parentSchema.getName(), true)
         .add(schema.name, new AbstractSchema());
-  }
-
-  public RelDataType getNamedType(SqlIdentifier typeName) {
-    if (typeName.equalsDeep(addressType.getSqlIdentifier(), Litmus.IGNORE)) {
-      return addressType;
-    } else {
-      return super.getNamedType(typeName);
-    }
   }
 
   private static List<RelCollation> deduceMonotonicity(
@@ -1092,7 +684,7 @@ public class MockCatalogReader extends CalciteCatalogReader {
     /**
      * A mock of ModifiableViewTable that can unwrap a mock RelOptTable.
      */
-    private static class MockModifiableViewTable extends ModifiableViewTable {
+    public static class MockModifiableViewTable extends ModifiableViewTable {
       private final RexNode constraint;
 
       MockModifiableViewTable(Type elementType, RelProtoDataType rowType,
@@ -1255,7 +847,7 @@ public class MockCatalogReader extends CalciteCatalogReader {
    * {@link org.apache.calcite.prepare.Prepare.PreparingTable} with dynamic record type.
    */
   public static class MockDynamicTable extends MockTable {
-    MockDynamicTable(MockCatalogReader catalogReader, String catalogName,
+    public MockDynamicTable(MockCatalogReader catalogReader, String catalogName,
         String schemaName, String name, boolean stream, double rowCount) {
       super(catalogReader, catalogName, schemaName, name, stream, rowCount,
           null, NullInitializerExpressionFactory.INSTANCE);
@@ -1378,140 +970,6 @@ public class MockCatalogReader extends CalciteCatalogReader {
     }
   }
 
-  /** Column having names with multiple parts. */
-  private static final class CompoundNameColumn {
-    final String first;
-    final String second;
-    final RelDataType type;
-
-    CompoundNameColumn(String first, String second, RelDataType type) {
-      this.first = first;
-      this.second = second;
-      this.type = type;
-    }
-
-    String getName() {
-      return (first.isEmpty() ? "" : ("\"" + first + "\"."))
-          + ("\"" + second + "\"");
-    }
-  }
-
-  /** ColumnResolver implementation that resolves CompoundNameColumn by simulating
-   *  Phoenix behaviors. */
-  private static final class CompoundNameColumnResolver implements ColumnResolver {
-    private final Map<String, Integer> nameMap = new HashMap<>();
-    private final Map<String, Map<String, Integer>> groupMap = new HashMap<>();
-    private final String defaultColumnGroup;
-
-    CompoundNameColumnResolver(
-        List<CompoundNameColumn> columns, String defaultColumnGroup) {
-      this.defaultColumnGroup = defaultColumnGroup;
-      for (Ord<CompoundNameColumn> column : Ord.zip(columns)) {
-        nameMap.put(column.e.getName(), column.i);
-        Map<String, Integer> subMap =
-            groupMap.computeIfAbsent(column.e.first, k -> new HashMap<>());
-        subMap.put(column.e.second, column.i);
-      }
-    }
-
-    @Override public List<Pair<RelDataTypeField, List<String>>> resolveColumn(
-        RelDataType rowType, RelDataTypeFactory typeFactory, List<String> names) {
-      List<Pair<RelDataTypeField, List<String>>> ret = new ArrayList<>();
-      if (names.size() >= 2) {
-        Map<String, Integer> subMap = groupMap.get(names.get(0));
-        if (subMap != null) {
-          Integer index = subMap.get(names.get(1));
-          if (index != null) {
-            ret.add(
-                new Pair<RelDataTypeField, List<String>>(
-                    rowType.getFieldList().get(index),
-                    names.subList(2, names.size())));
-          }
-        }
-      }
-
-      final String columnName = names.get(0);
-      final List<String> remainder = names.subList(1, names.size());
-      Integer index = nameMap.get(columnName);
-      if (index != null) {
-        ret.add(
-            new Pair<RelDataTypeField, List<String>>(
-                rowType.getFieldList().get(index), remainder));
-        return ret;
-      }
-
-      final List<String> priorityGroups = Arrays.asList("", defaultColumnGroup);
-      for (String group : priorityGroups) {
-        Map<String, Integer> subMap = groupMap.get(group);
-        if (subMap != null) {
-          index = subMap.get(columnName);
-          if (index != null) {
-            ret.add(
-                new Pair<RelDataTypeField, List<String>>(
-                    rowType.getFieldList().get(index), remainder));
-            return ret;
-          }
-        }
-      }
-      for (Map.Entry<String, Map<String, Integer>> entry : groupMap.entrySet()) {
-        if (priorityGroups.contains(entry.getKey())) {
-          continue;
-        }
-        index = entry.getValue().get(columnName);
-        if (index != null) {
-          ret.add(
-              new Pair<RelDataTypeField, List<String>>(
-                  rowType.getFieldList().get(index), remainder));
-        }
-      }
-
-      if (ret.isEmpty() && names.size() == 1) {
-        Map<String, Integer> subMap = groupMap.get(columnName);
-        if (subMap != null) {
-          List<Map.Entry<String, Integer>> entries =
-              new ArrayList<>(subMap.entrySet());
-          entries.sort((o1, o2) -> o1.getValue() - o2.getValue());
-          ret.add(
-              new Pair<RelDataTypeField, List<String>>(
-                  new RelDataTypeFieldImpl(
-                      columnName, -1,
-                      createStructType(
-                          rowType,
-                          typeFactory,
-                          entries)),
-                  remainder));
-        }
-      }
-
-      return ret;
-    }
-
-    private static RelDataType createStructType(
-        final RelDataType rowType,
-        RelDataTypeFactory typeFactory,
-        final List<Map.Entry<String, Integer>> entries) {
-      return typeFactory.createStructType(
-          StructKind.PEEK_FIELDS,
-          new AbstractList<RelDataType>() {
-            @Override public RelDataType get(int index) {
-              final int i = entries.get(index).getValue();
-              return rowType.getFieldList().get(i).getType();
-            }
-            @Override public int size() {
-              return entries.size();
-            }
-          },
-          new AbstractList<String>() {
-            @Override public String get(int index) {
-              return entries.get(index).getKey();
-            }
-            @Override public int size() {
-              return entries.size();
-            }
-          });
-    }
-  }
-
   /** Wrapper around a {@link MockTable}, giving it a {@link Table} interface.
    * You can get the {@code MockTable} by calling {@link #unwrap(Class)}. */
   private static class WrapperTable implements Table, Wrapper {
@@ -1584,167 +1042,6 @@ public class MockCatalogReader extends CalciteCatalogReader {
     }
   }
 
-  /** Default values for the "EMPDEFAULTS" table. */
-  private static class EmpInitializerExpressionFactory
-      extends NullInitializerExpressionFactory {
-    @Override public ColumnStrategy generationStrategy(RelOptTable table,
-        int iColumn) {
-      switch (iColumn) {
-      case 0:
-      case 1:
-      case 5:
-        return ColumnStrategy.DEFAULT;
-      default:
-        return super.generationStrategy(table, iColumn);
-      }
-    }
-
-    @Override public RexNode newColumnDefaultValue(RelOptTable table,
-        int iColumn, InitializerContext context) {
-      final RexBuilder rexBuilder = context.getRexBuilder();
-      final RelDataTypeFactory typeFactory = rexBuilder.getTypeFactory();
-      switch (iColumn) {
-      case 0:
-        return rexBuilder.makeExactLiteral(new BigDecimal(123),
-            typeFactory.createSqlType(SqlTypeName.INTEGER));
-      case 1:
-        return rexBuilder.makeLiteral("Bob");
-      case 5:
-        return rexBuilder.makeExactLiteral(new BigDecimal(555),
-            typeFactory.createSqlType(SqlTypeName.INTEGER));
-      default:
-        return rexBuilder.constantNull();
-      }
-    }
-  }
-
-  /** Types used during initialization. */
-  private class Fixture {
-    final RelDataType intType =
-        typeFactory.createSqlType(SqlTypeName.INTEGER);
-    final RelDataType intTypeNull =
-        typeFactory.createTypeWithNullability(intType, true);
-    final RelDataType varchar10Type =
-        typeFactory.createSqlType(SqlTypeName.VARCHAR, 10);
-    final RelDataType varchar10TypeNull =
-        typeFactory.createTypeWithNullability(varchar10Type, true);
-    final RelDataType varchar20Type =
-        typeFactory.createSqlType(SqlTypeName.VARCHAR, 20);
-    final RelDataType varchar20TypeNull =
-        typeFactory.createTypeWithNullability(varchar20Type, true);
-    final RelDataType timestampType =
-        typeFactory.createSqlType(SqlTypeName.TIMESTAMP);
-    final RelDataType timestampTypeNull =
-        typeFactory.createTypeWithNullability(timestampType, true);
-    final RelDataType dateType =
-        typeFactory.createSqlType(SqlTypeName.DATE);
-    final RelDataType booleanType =
-        typeFactory.createSqlType(SqlTypeName.BOOLEAN);
-    final RelDataType booleanTypeNull =
-        typeFactory.createTypeWithNullability(booleanType, true);
-    final RelDataType rectilinearCoordType =
-        typeFactory.builder()
-            .add("X", intType)
-            .add("Y", intType)
-            .build();
-    final RelDataType rectilinearPeekCoordType =
-        typeFactory.builder()
-            .add("X", intType)
-            .add("Y", intType)
-            .kind(StructKind.PEEK_FIELDS)
-            .build();
-    final RelDataType rectilinearPeekNoExpandCoordType =
-        typeFactory.builder()
-            .add("M", intType)
-            .add("SUB",
-                typeFactory.builder()
-                    .add("A", intType)
-                    .add("B", intType)
-                    .kind(StructKind.PEEK_FIELDS_NO_EXPAND)
-                    .build())
-            .kind(StructKind.PEEK_FIELDS_NO_EXPAND)
-            .build();
-    final RelDataType abRecordType =
-        typeFactory.builder()
-            .add("A", varchar10Type)
-            .add("B", varchar10Type)
-            .build();
-    final RelDataType skillRecordType =
-        typeFactory.builder()
-            .add("TYPE", varchar10Type)
-            .add("DESC", varchar20Type)
-            .add("OTHERS", abRecordType)
-            .build();
-    final RelDataType empRecordType =
-        typeFactory.builder()
-            .add("EMPNO", intType)
-            .add("ENAME", varchar10Type)
-            .add("DETAIL",
-                typeFactory.builder().add("SKILLS",
-                    typeFactory.createArrayType(skillRecordType, -1)).build())
-            .build();
-    final RelDataType empListType =
-        typeFactory.createArrayType(empRecordType, -1);
-
-    // TODO jvs 12-Feb-2005: register this canonical instance with type
-    // factory
-    final ObjectSqlType addressType =
-        new ObjectSqlType(SqlTypeName.STRUCTURED,
-            new SqlIdentifier("ADDRESS", SqlParserPos.ZERO),
-            false,
-            Arrays.asList(
-                new RelDataTypeFieldImpl("STREET", 0, varchar20Type),
-                new RelDataTypeFieldImpl("CITY", 1, varchar20Type),
-                new RelDataTypeFieldImpl("ZIP", 2, intType),
-                new RelDataTypeFieldImpl("STATE", 3, varchar20Type)),
-            RelDataTypeComparability.NONE);
-  }
-
-  /** To check whether
-   * {@link InitializerExpressionFactory#newColumnDefaultValue} is called.
-   *
-   * <p>If a column is in {@code defaultColumns}, returns 1 as the default
-   * value. */
-  public static class CountingFactory extends NullInitializerExpressionFactory {
-    static final ThreadLocal<AtomicInteger> THREAD_CALL_COUNT =
-        ThreadLocal.withInitial(AtomicInteger::new);
-
-    private final List<String> defaultColumns;
-
-    CountingFactory(List<String> defaultColumns) {
-      this.defaultColumns = ImmutableList.copyOf(defaultColumns);
-    }
-
-    @Override public ColumnStrategy generationStrategy(RelOptTable table,
-        int iColumn) {
-      final RelDataTypeField field =
-          table.getRowType().getFieldList().get(iColumn);
-      if (defaultColumns.contains(field.getName())) {
-        return ColumnStrategy.DEFAULT;
-      }
-      return super.generationStrategy(table, iColumn);
-    }
-
-    @Override public RexNode newColumnDefaultValue(RelOptTable table,
-        int iColumn, InitializerContext context) {
-      THREAD_CALL_COUNT.get().incrementAndGet();
-      final RelDataTypeField field =
-          table.getRowType().getFieldList().get(iColumn);
-      if (defaultColumns.contains(field.getName())) {
-        final RexBuilder rexBuilder = context.getRexBuilder();
-        return rexBuilder.makeExactLiteral(BigDecimal.ONE);
-      }
-      return super.newColumnDefaultValue(table, iColumn, context);
-    }
-
-    @Override public RexNode newAttributeInitializer(RelDataType type,
-        SqlFunction constructor, int iAttribute,
-        List<RexNode> constructorArgs, InitializerContext context) {
-      THREAD_CALL_COUNT.get().incrementAndGet();
-      return super.newAttributeInitializer(type, constructor, iAttribute,
-         constructorArgs, context);
-    }
-  }
 }
 
 // End MockCatalogReader.java
